New bot control loop in a separate subprocess
This should be much faster, but still needs more
debugging, and is currently a bit less debuggable.

One subprocess tries to stay in sync with the tablet
event rate, and updates the motors. It uses queues to
accept updated goal images and deliver status updates
to another process, which does the slower image operations.